requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/requirements.in
#
amqp==5.1.1
    # via kombu
anyascii==0.3.1
    # via wagtail
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
attrs==21.4.0
    # via jsonschema
beautifulsoup4==4.9.3
    # via wagtail
billiard==3.6.4.0
    # via celery
celery[redis]==5.2.7
    # via -r requirements/requirements.in
celery-progress==0.1.2
    # via -r requirements/requirements.in
certifi==2022.5.18.1
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==37.0.2
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
deprecated==1.2.13
    # via redis
dj-stripe==2.6.1
    # via -r requirements/requirements.in
django==3.2.14
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
    #   django-allauth
    #   django-anymail
    #   django-hijack
    #   djangorestframework
    #   jsonfield
django-allauth==0.50.0
    # via -r requirements/requirements.in
django-anymail[mailgun]==8.6
    # via -r requirements/requirements.in
django-filter==21.1
    # via wagtail
django-hijack==3.2.1
    # via -r requirements/requirements.in
django-modelcluster==6.0
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-taggit==2.1.0
    # via wagtail
django-treebeard==4.5.1
    # via wagtail
djangorestframework==3.13.1
    # via
    #   -r requirements/requirements.in
    #   drf-spectacular
djangorestframework-api-key==2.2.0
    # via -r requirements/requirements.in
draftjs-exporter==2.1.7
    # via wagtail
drf-spectacular==0.22.1
    # via -r requirements/requirements.in
et-xmlfile==1.1.0
    # via openpyxl
html5lib==1.1
    # via wagtail
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonfield==3.1.0
    # via dj-stripe
jsonschema==4.6.0
    # via drf-spectacular
kombu==5.2.4
    # via celery
l18n==2021.3
    # via wagtail
oauthlib==3.2.0
    # via requests-oauthlib
openpyxl==3.0.10
    # via tablib
packaging==21.3
    # via redis
pillow==9.1.1
    # via wagtail
prompt-toolkit==3.0.29
    # via click-repl
psycopg2-binary==2.9.3
    # via -r requirements/requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.4.0
    # via django-allauth
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python3-openid==3.2.0
    # via django-allauth
pytz==2022.1
    # via
    #   celery
    #   django
    #   djangorestframework
pyyaml==6.0
    # via drf-spectacular
redis==4.3.1
    # via celery
requests==2.27.1
    # via
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
    #   stripe
requests-oauthlib==1.3.1
    # via django-allauth
sentry-sdk==1.5.12
    # via -r requirements/requirements.in
six==1.16.0
    # via click-repl
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
stripe==3.2.0
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
tablib[xls,xlsx]==3.2.1
    # via wagtail
telepath==0.2
    # via wagtail
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.9
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wagtail==3.0
    # via -r requirements/requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via html5lib
willow==1.4.1
    # via wagtail
wrapt==1.14.1
    # via deprecated
xlrd==2.0.1
    # via tablib
xlsxwriter==3.0.3
    # via wagtail
xlwt==1.3.0
    # via tablib
